+++ MiMap.Common/Net/MapConfigPacket.cs
@@ -4,7 +4,7 @@ namespace MiMap.Common.Net
 {
     public class MapConfigPacket : IPacket
     {
-        public byte Id { get; } = Protocol.MapConfig;
+        public int Id { get; } = Protocol.MapConfig;
 
         public MiMapConfig Config { get; set; }
     }
