+++ MiMap.Common/Net/Protocol.cs
@@ -2,12 +2,12 @@
 {
     public class Protocol
     {
-        public static readonly byte Null = 0;
-        public static readonly byte MapConfig = 1;
-        public static readonly byte TileUpdate = 2;
-        public static readonly byte LevelMeta = 3;
-        public static readonly byte ListPlayers = 4;
-        public static readonly byte TileSubscribe = 5;
+        public static readonly int Null = 0;
+        public static readonly int MapConfig = 1;
+        public static readonly int TileUpdate = 2;
+        public static readonly int LevelMeta = 3;
+        public static readonly int ListPlayers = 4;
+        public static readonly int TileSubscribe = 5;
     }
 }
